Rename model module -> user_model
