+++ src/lcm/__init__.py
@@ -7,6 +7,6 @@ except ImportError:
 
 from lcm import mark
 from lcm.grids import DiscreteGrid, LinspaceGrid, LogspaceGrid
-from lcm.model import Model
+from lcm.user_model import Model
 
 __all__ = ["mark", "Model", "LinspaceGrid", "LogspaceGrid", "DiscreteGrid"]
